+++ CHANGELOG.md
@@ -8,17 +8,22 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## Version 1.1.0 - Unreleased
 
 ### Added
+* New method: `ub.timeparse` can parse the result of `ub.timestamp` into a
+  `datetime` object. Can optionally use `dateutil.parser.parse` under the hood.
 * `ub.cmd` now has a `system` argument for modularity with `os.system`.
 * `ub.cmd` now accepts a `timeout` argument (tee support is pending).
 * `ub.JobPool` now contains a protected `_prog` variable allowing the user
   finer-grained progress controls.
 * `ub.JobPool` now contains a convenience method `join` that executes all jobs
   and returns a list of results.
-* `ub.timestamp` can now accept a datetime object as an argument, and will return 
+* `ub.timestamp` can now accept a `datetime` object as an argument, and will return 
    the timestamp for that object.
 * `ub.Path.ls` a convenience function that aliases `list(path.iterdir())`.
 * `ub.Path.walk` to wrap `os.walk`. 
-
+* `ub.CacheStamp` will now check the mtime and size to quickly check if the products
+  have changed and force expiration.
+* `ub.CacheStamp` now takes an expires keyword arg, which will keep the cache valid 
+  only for the specified amount of time.
 
 ### Changed
 * Register `pathlib.Path` with `ub.repr2`
@@ -29,6 +34,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed
 * `ub.hash_data` now recognizes subclasses of registered types.
 * `ub.timestamp()` has been outputting incorrect (negated) UTC offsets. This is now fixed.
+* `ub.timestamp()` now works correctly when the year has less than 4 digits.
 
 ### Changed
 * The `ubelt.util_hash.HashableExtensions` implementation was updated to use